requirements/production.txt
# Dependencies related to deploying rugbystat
-r base.txt

# WSGI Handler
gevent==1.2.0
gunicorn==19.6.0

# Caching
hiredis==0.2.0
django-redis-cache==1.7.1

# Django-secure
django-secure==1.0.1

# App Monitoring
newrelic==2.76.0.55
